requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==23.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.8
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
